# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Go LineLess API specification — Ahmedabad Pilot
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/me:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's active subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions:
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe to a plan (cancels existing active subscription)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ─── NOTIFICATIONS ──────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user or runner
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read (with ownership check)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PAYMENTS ──────────────────────────────────────────────────────────────
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a Razorpay payment order for a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [taskId]
              properties:
                taskId:
                  type: integer
      responses:
        "200":
          description: Payment order created (or cash method info)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /subscriptions/create-order:
    post:
      operationId: createSubscriptionOrder
      tags: [subscriptions]
      summary: Create a Razorpay payment order for a subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId:
                  type: string
                billingCycle:
                  type: string
                  enum: [monthly, yearly]
                  default: monthly
      responses:
        "200":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /payments/webhook:
    post:
      operationId: handlePaymentWebhook
      tags: [payments]
      summary: Receive Razorpay webhook events (payment.captured, payment.failed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        "401":
          $ref: "#/components/responses/ErrorResponse"

  # ─── UPLOAD ─────────────────────────────────────────────────────────────────
...
  /admin/subscriptions:
    get:
      operationId: listAdminSubscriptions
      tags: [admin]
      summary: List all subscriptions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
              detail:
                type: string
              code:
                type: string
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, planId, planName, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: string
        planName:
          type: string
        billingCycle:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [active, cancelled, expired]
        tasksPerMonth:
          type: integer
          nullable: true
        tasksUsed:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        amount:
          type: number
        createdAt:
          type: string
    SubscriptionInput:
      type: object
      required: [planId, billingCycle]
      properties:
        planId:
          type: string
        billingCycle:
          type: string
          enum: [monthly, yearly]

    # ─── NOTIFICATION SCHEMAS ─────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        runnerId:
          type: integer
          nullable: true
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        taskId:
          type: integer
          nullable: true
        createdAt:
          type: string

    # ─── REVIEW SCHEMAS ──────────────────────────────────────────────────────
...
    PaymentOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          nullable: true
          description: Razorpay order ID (null for cash or mock orders)
        amount:
          type: integer
          description: Amount in paise
        currency:
          type: string
        receipt:
          type: string
          nullable: true
        keyId:
          type: string
          nullable: true
          description: Razorpay key ID for frontend checkout
        mock:
          type: boolean
          description: True when Razorpay is not configured (dev mode)
        method:
          type: string
          description: Payment method (online or cash)
        price:
          type: number
          description: Original task price

    # ─── OPERATIONS SCHEMAS ──────────────────────────────────────────────────
...